add order and order complete page
creazione ordine + nuovo carrello + pagina review ordine

+++ cart.php
@@ -13,19 +13,30 @@ $check_idChart = "SELECT * FROM cart_info where id_cart='$id_cart'";
 $run1 = mysqli_query($mysqli, $check_idChart);
 $total_price = 0;
 
-while ($data = $run1->fetch_assoc()) {
+
+$rowCount = mysqli_num_rows($run1);
+
+if($rowCount != 0){
+    while ($data = $run1->fetch_assoc()) {
     $id_product = $data['id_product'];
     $main->setContent("image", $data['image']);
     $main->setContent("name", $data['name']);
     $quantity = $data['quantity'];
     $main->setContent("price", $data['price'] . "€");
     $price = $data['price'];
-    $main->setContent("quantity", $data['quantity']);
+    $main->setContent("quantityInput", "<input type='number' id='quantity-value1' min='0' value='$quantity'>");
     $total_product_price = $price * $quantity;
     $main->setContent("productPrice", "$total_product_price" . "€");
     $main->setContent("delete", "<a href='deleteItemCart.php?idProduct=$id_product&idCart=$id_cart' class='uk-icon-button uk-icon-times-circle'></a>");
     $total_price = $total_price + $total_product_price;
 }
+    $main->setContent("checkout", "<a href='checkout.php' class='uk-button uk-button-small idz-button-white uk-width-1-1'>Proceed to Checkout</a>");
+    $main->setContent("update", "<a href='' class='uk-button uk-button-small idz-button-grey uk-margin-small-top'>Update Cart</a>");
+
+} else {
+    $main->setContent("name", "There isn't any product in your cart");
+}
+
 
 $main->setContent("totalPrice", "$total_price" . "€");
